Create game endpoints

+++ src/main/java/com/boardcamp/api/dtos/GameDTO.java
@@ -18,5 +18,5 @@ public class GameDTO {
     private Integer stockTotal;
 
     @NotBlank(message = "Price is required")
-    private Integer pricePerDay;
+    private Double pricePerDay;
 }
